Modified setup
Moved HeroAI out of widgets, as it causes crashes on load currently.

Added au3 injector into main directory as this is my preferred method of injection and seems nice to have as an option for all.

Finally, I updated combat.py to use a dictionary of skillID : lambda pairs for custom function evaluations instead of a large if tree.