Utility Overhaul
Huge update, changed the main engine and debugging info to utilize a utility evaluation of functions for more robust behavior.

Requires further testing for full stability.

My goal was to maintain old behavior as closely as possible with this update for now.

ONLY IMPLEMENTED FOR SKILLS, MOVEMENT AND OTHER CHOICES ARE MADE USING DECISION TREES STILL.