Improve email retrieval with UUID check
This fixes some issues introduced with the uuid formatting as it was to simple.

The issue was GetAccountEmail would return "uuid_0_0_0_0" until the email is available, which is breaking all kind of logic. If the uuid is empty (0,0,0,0) it should return an empty string to indicate there is no value available. Only if the uuid is not empty and the email isn't available it should create that uuid string.